# docs/swagger/index.yml
openapi: 3.0.3
info:
  title: online-shop-express
  version: 1.0.0
  contact: {}
servers:
  - url: 127.0.0.1
paths:
...
  /api/v1/category/6527164e370e0ccef394e674:
    get:
      tags:
        - category-service
      summary: read-category-by-Id
      description: read-category-by-Id
      operationId: readCategoryById
      parameters:
        - name: page
          in: query
          schema:
            type: string
            example: '1'
      responses:
        '200':
          description: ''
    patch:
      tags:
        - category-service
      summary: update-category
      description: update-category
      operationId: updateCategory
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: New Category Name As C2
                name:
                  type: string
                  example: C2 updated3
            examples:
              update-category:
                value:
                  description: New Category Name As C2
                  name: C2 updated3
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
  /api/v1/category/653680915b54be8eb91c59fe:
    delete:
      tags:
        - category-service
      summary: disable-category
      description: disable-category
      operationId: disableCategory
      requestBody:
        content:
          application/json:
            examples:
              disable-category:
                value: ''
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
  /api/v1/category/:
    post:
      tags:
        - category-service
      summary: create-category
      description: create-category
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: New Category
                name:
                  type: string
                  example: C 04
                thumbnail:
                  type: string
                  example: 1699220385727-banana.png
            examples:
              create-category:
                value:
                  description: New Category
                  name: C 04
                  thumbnail: 1699220385727-banana.png
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
  /api/v1/category/thumbnail/653832ad2829cf39683badda:
    patch:
      tags:
        - category-service
      summary: update-category-thumbnail
      description: update-category-thumbnail
      operationId: updateCategoryThumbnail
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                thumbnail:
                  type: string
                  example: 1698182343185-banana.png
            examples:
              update-category-thumbnail:
                value:
                  thumbnail: 1698182343185-banana.png
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
  /api/v1/category/enable/653680915b54be8eb91c59fe:
    post:
      tags:
        - category-service
      summary: enable-category
      description: enable-category
      operationId: enableCategory
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
  /api/v1/category/thumbnail/653680915b54be8eb91c59fe:
    delete:
      tags:
        - category-service
      summary: delete-category-thumbnail
      description: delete-category-thumbnail
      operationId: deleteCategoryThumbnail
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
  /api/v1/cart:
    get:
      tags:
        - cart-service
      summary: read-cart
      description: read-cart
      operationId: readCart
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...